agent-video/requirements.txt
functions-framework==3.*
google-cloud-storage>=2.0.0
google-cloud-firestore>=2.11.0
google-genai>=0.8.0
requests>=2.31.0
google-auth>=2.23.0